requirements.txt
python-dotenv
rich
langchain
langgraph
openai
langchain-openai
